requirements.txt
flask==3.0.0
firebase-admin==6.3.0
python-dotenv==1.0.0
PyJWT==2.8.0
pytest==7.4.3
requests==2.31.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dateutil==2.8.2
yfinance==0.2.36
pyrebase4==4.7.1